// common/types/Class.ts
export enum Class {
    monk = "monk",
    demon_hunter = "demon-hunter",
    death_knight = "death-knight",
    druid = "druid",
    hunter = "hunter",
    mage = "mage",
    paladin = "paladin",
    priest = "priest",
    rogue = "rogue",
    shaman = "shaman",
    warlock = "warlock",
    warrior = "warrior",
}